Trait notion::command::Command

pub(crate) trait Command: Sized {
    type Args: DeserializeOwned;

    const USAGE: &'static str;

    fn help() -> Self;

    fn parse(notion: Notion, args: Self::Args) -> Fallible<Self>;

    fn run(self, session: &mut Session) -> Fallible<()>;

    fn go(notion: Notion, session: &mut Session) -> Fallible<()> { ... }
}

A Notion command.

Associated Types

type Args: DeserializeOwned

The intermediate type Docopt should deserialize the parsed command into.

Associated Constants

const USAGE: &'static str

The full usage documentation for this command. This can contain leading and trailing whitespace, which will be trimmed before printing to the console.

Required Methods

fn help() -> Self

Produces a variant of this type representing the notion <command> --help option.

fn parse(notion: Notion, args: Self::Args) -> Fallible<Self>

Parses the intermediate deserialized arguments into the full command.

fn run(self, session: &mut Session) -> Fallible<()>

Executes the command. Returns Ok(true) if the process should return 0, Ok(false) if the process should return 1, and Err(e) if the process should return e.exit_code().

Provided Methods

fn go(notion: Notion, session: &mut Session) -> Fallible<()>

Top-level convenience method for taking a Notion invocation and executing this command with the arguments taken from the Notion invocation.